refactor: Strategy, Factory 패턴 예제를 main.go로 실행하도록 변경
- example_test.go 제거하고 각 패턴 폴더에 main.go 추가
- 패키지 소스 파일을 pkg/ 하위로 이동하여 go run main.go 실행 가능
- README.md 파일 구성 및 실행 방법 업데이트
