+++ golang/pattern/factory/pkg/email_service.go
@@ -1,4 +1,4 @@
-package factory
+package pkg
 
 import (
 	"context"
