+++ golang/pattern/factory/pkg/provider.go
@@ -1,4 +1,4 @@
-package factory
+package pkg
 
 import (
 	"context"
