+++ golang/pattern/factory/pkg/user_preference_store.go
@@ -1,4 +1,4 @@
-package factory
+package pkg
 
 import (
 	"context"
