+++ golang/pattern/strategy/README.md
@@ -53,12 +53,13 @@ classDiagram
 
 | 파일 | 설명 |
 |------|------|
-| `domain.go` | `PaymentStrategy` 인터페이스 및 `PaymentResult` 타입 정의 |
-| `credit_card_service.go` | 신용카드 결제 서비스 구현체 |
-| `kakao_pay_service.go` | 카카오페이 결제 서비스 구현체 |
-| `naver_pay_service.go` | 네이버페이 결제 서비스 구현체 |
-| `payment_usecase.go` | 결제 유스케이스 (Context) |
-| `helper.go` | 헬퍼 함수 |
+| `main.go` | 실행 진입점 (예제 코드) |
+| `pkg/domain.go` | `PaymentStrategy` 인터페이스 및 `PaymentResult` 타입 정의 |
+| `pkg/credit_card_service.go` | 신용카드 결제 서비스 구현체 |
+| `pkg/kakao_pay_service.go` | 카카오페이 결제 서비스 구현체 |
+| `pkg/naver_pay_service.go` | 네이버페이 결제 서비스 구현체 |
+| `pkg/payment_usecase.go` | 결제 유스케이스 (Context) |
+| `pkg/helper.go` | 헬퍼 함수 |
 
 ## 핵심 구성요소
 
@@ -136,8 +137,8 @@ func TestCreditCardService(t *testing.T) {
 }
 ```
 
-## 테스트 실행
+## 실행
 
 ```bash
-go test -v -run TestStrategyPattern ./...
+go run main.go
 ```
