+++ golang/pattern/strategy/pkg/credit_card_service.go
@@ -1,4 +1,4 @@
-package strategy
+package pkg
 
 import (
 	"context"
