+++ golang/pattern/strategy/pkg/domain.go
@@ -1,4 +1,4 @@
-package strategy
+package pkg
 
 import "context"
