+++ golang/pattern/strategy/pkg/helper.go
@@ -1,4 +1,4 @@
-package strategy
+package pkg
 
 func generateID() string {
 	return "12345678"
