+++ golang/pattern/strategy/pkg/naver_pay_service.go
@@ -1,4 +1,4 @@
-package strategy
+package pkg
 
 import (
 	"context"
