+++ golang/pattern/strategy/pkg/payment_usecase.go
@@ -1,4 +1,4 @@
-package strategy
+package pkg
 
 import (
 	"context"
